refactor(clinker-core): split plan/execution.rs — types layer hub
Convert plan/execution.rs into a plan/execution/ directory whose mod.rs
holds the shared type layer — PlanNode (and its payload structs),
PlanEdge, the ExecutionPlanDag struct, NodeExecutionReqs, and PlanError —
and re-exports each submodule so every crate::plan::execution::* path
external and internal consumers use resolves unchanged.

The two cross-submodule render/classification helpers (combine_strategy_tag,
arbitration_class/ArbitrationClass) live here at pub(super) so the explain
and scheduling submodules share one definition.

Pure relocation: no symbol renames, no public-API change.

Closes #249
